Maximum Number of Balloons
One of the Tesla questions. Here we have to check whether the string contains character to form the word balloon. So let's say we have 2 instances of 'b', 3 instances of 'a', 4 instances of 'l', 5 instances of 'o', and 3 instances of 'n'. The maximum number of words that can be formed will be by finding out the minimum number of instances of a word in the string. Using the min function we can find the minimum of each letter in the word and simply returning it.